requirement.txt
astroid==2.0.4
certifi==2018.8.24
chardet==3.0.4
defusedxml==0.5.0
Django==2.2
django-allauth==0.37.1
django-cors-headers==2.4.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
gunicorn==19.7.1
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==2.1.0
pylint==2.1.1
python3-openid==3.1.0
pytz==2018.5
requests==2.21.0
requests-oauthlib==1.0.0
six==1.11.0
sqlparse==0.3.0
typed-ast==1.1.0
urllib3==1.24.2
whitenoise==3.3.1
wrapt==1.10.11